rewrite load-user-plugins to use new compound plugin api from secret-stack@6.1
